pip_requirements.txt
blinker
flask
flup
python-cjson
raven
